+++ src/FSharp.HotChocolate.Tests/Nullability.fs
@@ -478,7 +478,7 @@ type Query() =
 
 let builder =
     ServiceCollection()
-        .AddGraphQLServer(disableCostAnalyzer = true)
+        .AddGraphQLServer()
         .AddQueryType<Query>()
         .AddFSharpSupport()
         .AddTypeConverter<Uri, string>(string<Uri>)