+++ src/FSharp.HotChocolate.Tests/Nullability_GlobalIdentification.fs
@@ -101,7 +101,7 @@ type Query() =
 
 let builder =
     ServiceCollection()
-        .AddGraphQLServer(disableCostAnalyzer = true)
+        .AddGraphQLServer()
         .AddQueryType<Query>()
         .AddFSharpSupport()
         .AddGlobalObjectIdentification()