+++ src/FSharp.HotChocolate.Tests/UnionsAsUnions.fs
@@ -87,7 +87,7 @@ type Query() =
 
 let builder =
     ServiceCollection()
-        .AddGraphQLServer(disableCostAnalyzer = true)
+        .AddGraphQLServer()
         .AddQueryType<Query>()
         .AddFSharpSupport()
         .AddType<ADescriptor>()